全てのjavascriptファイルをtypescriptに移行しました(型要修正)
