+++ app/views/posts/detail.scala.html
@@ -121,5 +121,5 @@
 		</section>
 	</section>
 </div>
-<script src="@routes.Assets.versioned("javascripts/posts/detail.js")"></script>
+<script type="module" src="@routes.Assets.versioned("javascripts/posts_detail.js")"></script>
 }
