+++ app/views/posts/index.scala.html
@@ -122,5 +122,5 @@ UserOptRequestHeader,flash: Flash)
 			</section>
 		</section>
 	</div>
-	<script type="module" src="@routes.Assets.versioned("javascripts/post_index.js")"></script>
+	<script type="module" src="@routes.Assets.versioned("javascripts/posts_index.js")"></script>
 	}
