+++ app/views/ranking/index.scala.html
@@ -87,5 +87,5 @@
 		</section>
 	</section>
 </div>
-<script src="@routes.Assets.versioned("javascripts/ranking/index.js")"></script>
+<script type="module" src="@routes.Assets.versioned("javascripts/ranking_index.js")"></script>
 }
