+++ app/views/users/detail.scala.html
@@ -110,6 +110,6 @@
 			}
 		</section>
 </div>
-<script src="@routes.Assets.versioned("javascripts/users/detail.js")" type="text/javascript"></script>
+<script type="module" src="@routes.Assets.versioned("javascripts/users_detail.js")" type="text/javascript"></script>
 </div>
 }
