+++ app/views/users/edit.scala.html
@@ -53,5 +53,5 @@ UserRequestHeader,flash: Flash)
 		}
 	</section>
 </div>
-<script src="@routes.Assets.versioned("javascripts/users/edit.js")"></script>
+<script type="module" src="@routes.Assets.versioned("javascripts/users_edit.js")"></script>
 }
