+++ app/views/users/sign_in.scala.html
@@ -34,5 +34,5 @@
 		}
 	</section>
 </div>
-<script src="@routes.Assets.versioned("javascripts/users/signIn.js")"></script>
+<script type="module" src="@routes.Assets.versioned("javascripts/users_signIn.js")"></script>
 }
