+++ frontend/src/typescripts/users/signIn.ts
@@ -1,0 +1,5 @@
+export { }
+const $errorNeedSignIn = <HTMLElement>document.getElementById("errorNeedSignIn");
+if (location.search == "?=needSignIn") {
+	$errorNeedSignIn.classList.remove("display-none");
+}
