+++ public/javascripts/posts_detail.js
@@ -1,0 +1,1 @@
+import{i as p,b as I,r as l,c as g,e as f,a as E,d as L}from"./main.js";const i=document.getElementById("commentInsertButton"),T=document.getElementById("commentInsertForm"),o=document.getElementById("content"),k=document.getElementById("commentList"),r=document.getElementById("commentCount");let a=parseInt(r.innerText);const m=T.children[0].value,b=document.getElementById("postId").value;i.setAttribute("disabled","true");o.addEventListener("input",s=>{if(!(s.target instanceof HTMLTextAreaElement))return;s.target.value.replace(/\r?\n/g,"").replace(/\s+/g,"")?i.removeAttribute("disabled"):i.setAttribute("disabled","true")});i.addEventListener("click",s=>{let e=o.value;p.post("/comment",{postId:b,content:e},{headers:{"Content-Type":"application/json","Csrf-Token":m}}).then(t=>{I(k,t.data),l(o),o.value="",r.innerText==""?(a=1,r.innerText=a.toFixed()):(a+=1,r.innerText=a.toFixed())}).catch(t=>{t.response.status==401&&(location.href="/users/sign-in?=needSignIn"),l(o);let c=t.response.data.content[0];g(f(c),o)})});const v=document.getElementsByClassName("likeInfo");Array.from(v).map(s=>{const e=s.children[0],t=s.children[1],c=n=>{t.innerText=n.data,e.classList.remove("like-btn"),e.classList.add("unlike-btn")},u=n=>{n.data==0?t.innerText="":t.innerText=n.data,e.classList.remove("unlike-btn"),e.classList.add("like-btn")};e.addEventListener("click",n=>{if(!(n.target instanceof HTMLElement))return;const d=n.target.dataset.postId;n.target.classList.contains("like-btn")?E(d,m,c):L(d,m,u)})});
