+++ public/javascripts/posts_index.css
