+++ public/javascripts/posts_index.js
