+++ public/javascripts/ranking_index.js
@@ -1,0 +1,1 @@
+import{a as u,d as m}from"./main.js";const L=new URL(location.href),f=L.search,c=decodeURI(f),k=c.indexOf("department"),v=c.indexOf("&"),a=c.substring(k+10+1,v),s=document.getElementsByClassName("nav-link");Array.from(s).forEach(n=>n.classList.remove("active"));a=="\u30D5\u30ED\u30F3\u30C8\u30A8\u30F3\u30C9"?s[1].classList.add("active"):a=="\u30D0\u30C3\u30AF\u30A8\u30F3\u30C9"?s[2].classList.add("active"):a=="\u30AF\u30E9\u30A6\u30C9"?s[3].classList.add("active"):s[0].classList.add("active");const o=document.getElementById("signOutForm").children[0].value,h=document.getElementsByClassName("likeInfo");Array.from(h).map(n=>{const t=n.children[0],i=n.children[1],d=e=>{i.innerText=e.data,t.classList.remove("like-btn"),t.classList.add("unlike-btn")},l=e=>{i.innerText=e.data,t.classList.remove("unlike-btn"),t.classList.add("like-btn")};t.addEventListener("click",e=>{if(!(e.target instanceof HTMLElement))return;const r=e.target.dataset.postId;e.target.classList.contains("like-btn")?u(r,o,d):m(r,o,l)})});
