--- public/javascripts/users/signIn.js
@@ -1,6 +1,0 @@
-(function () {
-  const $errorNeedSignIn = document.getElementById("errorNeedSignIn");
-  if (location.search == "?=needSignIn") {
-    $errorNeedSignIn.classList.remove("display-none");
-  }
-})();
