+++ public/javascripts/users_detail.js
@@ -1,0 +1,1 @@
+import{a as y,d as B,i as C,f as h}from"./main.js";const E=document.getElementsByClassName("deleteForm");[...E].map(e=>{e.children[1].addEventListener("click",n=>{confirm("\u672C\u5F53\u306B\u524A\u9664\u3057\u3066\u3088\u308D\u3057\u3044\u3067\u3059\u304B\uFF1F")&&e.submit()})});const D=document.getElementsByClassName("likeInfo"),u=E[0].children[0].value;Array.from(D).map(e=>{const t=e.children[0],n=e.children[1],o=s=>{n.innerText=s.data,t.classList.remove("like-btn"),t.classList.add("unlike-btn")},i=s=>{n.innerText=s.data,t.classList.remove("unlike-btn"),t.classList.add("like-btn")};t.addEventListener("click",s=>{if(!(s.target instanceof HTMLElement))return;const r=s.target.dataset.postId;s.target.classList.contains("like-btn")?y(r,u,o):B(r,u,i)})});const b=document.getElementsByClassName("editFormGenerateButton"),k=document.getElementsByClassName("card-textarea");Array.from(b).map(e=>e.addEventListener("click",t=>{if(!(t.target instanceof HTMLElement))return;const n=t.target.dataset.postId,o=M(n),i=o.children[1],s=o.children[0],r=i.children[0];x(i,s),p(r,m(s.innerText))}));Array.from(k).map(e=>{const t=e.parentNode,n=t.previousElementSibling,o=t.parentNode.nextElementSibling,i=e.nextElementSibling.firstElementChild.firstElementChild,s=parseInt(t.parentNode.dataset.postId);e.addEventListener("focus",()=>{o.classList.add("display-none"),i.addEventListener("click",r=>{r.stopPropagation(),v("jsPostErrorMsgDiv");const a=e.value;C.post("/posts/update",{postId:s,content:a,postedAt:Date.now()},{headers:{"Csrf-Token":u}}).then(c=>{const d=c.data,l=t.parentNode.previousElementSibling.children[0].lastElementChild;n.innerText=d.content,l.innerText=h(d.postedAt),g(t,n),o.classList.remove("display-none")}).catch(c=>{const l=c.response.data.errors["obj.content"][0].msg[0],f=t.parentNode;F(f,l),p(e,m(a))})})},{once:!0}),e.addEventListener("blur",r=>{const a=r.relatedTarget;if(a==null||a.name!="editDecideButton")if(confirm("\u7DE8\u96C6\u5185\u5BB9\u3092\u7834\u68C4\u3057\u3066\u3082\u3088\u308D\u3057\u3044\u3067\u3057\u3087\u3046\u304B\uFF1F")){v("jsPostErrorMsgDiv");const c=n.innerText;e.value=c,g(t,n),o.classList.remove("display-none")}else setTimeout(()=>{p(e,m(e.value))},10)})});const M=e=>{const t=document.querySelectorAll("section.card-body");let n=null;return Array.from(t).map(o=>{o.dataset.postId==e&&(n=o)}),n},x=(e,t)=>{t.classList.add("display-none"),e.classList.remove("display-none")},g=(e,t)=>{e.classList.add("display-none"),t.classList.remove("display-none")},m=e=>e.length,p=(e,t)=>{e.focus(),e.setSelectionRange(t,t)},F=(e,t)=>{const n=S(t);e.prepend(n)},S=e=>{const t=document.createElement("div");return t.innerText=e,t.style.padding="16px",t.classList.add("error-message"),t.id="jsPostErrorMsgDiv",t},v=e=>{const t=document.getElementById(e);t!=null&&t.remove()};
