+++ public/javascripts/users_edit.js
@@ -1,0 +1,1 @@
+const n=document.getElementById("profileImgInput"),i=document.getElementById("profileImgError"),o=document.getElementById("preview"),s=o.children.item(0),u=document.getElementById("updateButton"),d=()=>{i.innerText="",u.disabled=!1},c=t=>t.size>2097152?(i.innerText="\u30D5\u30A1\u30A4\u30EB\u30B5\u30A4\u30BA\u304C\u5927\u304D\u3059\u304E\u307E\u3059\u30022MB\u4EE5\u5185\u306B\u3057\u3066\u304F\u3060\u3055\u3044",u.disabled=!0,!1):!0,a=t=>{const e=new FileReader;e.onload=r=>{const l=r.target.result;s.src=l},e.readAsDataURL(t)},m=()=>{d();const t=n.files;Array.from(t).map(e=>{c(e)==!0&&a(e)})};n.addEventListener("change",()=>{m()});
