+++ public/javascripts/users_signIn.js
@@ -1,0 +1,1 @@
+const e=document.getElementById("errorNeedSignIn");location.search=="?=needSignIn"&&e.classList.remove("display-none");
